+++ pos_app/src/main/java/com/increff/pos/model/data/InventoryReportData.java
@@ -11,5 +11,5 @@ public class InventoryReportData {
     private String brand;
     private String category;
     private String barcode;
-    private Long qty;
+    private Integer qty;
 }
